pyproject.toml
[project]
name = "lm-reward-jax"
version = "0.1.0"
description = "Official implementation of MVR: Multi-view Video Reward Shaping for Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = "MIT"
authors = [
    { name = "Lirui Luo" },
    { name = "Guoxi Zhang" },
    { name = "Hongming Xu" },
    { name = "Yaodong Yang" },
    { name = "Cong Fang" },
    { name = "Qing Li" },
]
dependencies = [
    "accelerate>=1.9.0",
    "diffusers>=0.34.0",
    "einops>=0.8.1",
    "humanoid-bench",
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "imageio[ffmpeg,pyav]>=2.37.0",
    "isort>=6.0.1",
    "jax[cuda12]==0.5.0",
    "matplotlib>=3.10.3",
    "mediapy>=1.2.4",
    "metaworld",
    "open-clip-torch>=2.32.0",
    "sbx-rl>=0.20.0",
    "sentence-transformers>=5.0.0",
    "tabulate>=0.9.0",
    "tbparse>=0.0.9",
    "tensorboard>=2.19.0",
    "torch==2.7.1",
    "torchaudio==2.7.1",
    "torchvision==0.22.1",
    "tritonclient[all]>=2.59.0",
    "voltron-robotics>=0.0.1",
    "wandb>=0.21.0",
]

[project.urls]
Homepage = "https://mvr-rl.github.io/"
Paper = "https://openreview.net/forum?id=7lw6s9ELfr"

[tool.uv.sources]
metaworld = { git = "https://github.com/Farama-Foundation/Metaworld.git" }
humanoid-bench = { git = "https://github.com/liruiluo/humanoid-bench.git" }
